Move devs, wins, and scripts/webs into cases; unify line endings
Restructuring: cases/ is now the single top-level container for all
script categories.
- devs/ -> cases/devs/
- wins/ -> cases/wins/
- scripts/webs/ -> cases/webs/ (merged with the existing cases/webs/
  folder rather than nested separately - lucky/pinterest/openwebs/
  mp4to3/youtube all live together there now)
- scripts/README.md -> cases/README.md

scripts/ now only holds hello.py, which wasn't part of what was
asked to move.

Updated everything that referenced the old paths: pyproject.toml
console-script entries, test imports, the taskbar setup script's
REPO_ROOT calculation (one directory deeper now), the three taskbar
.bat wrappers' relative paths to .venv, the README's per-folder
sections (now cases/devs, cases/maps, cases/pictures, cases/webs,
cases/wins, scripts), and the top-level README's link to it.

Also re-ran the taskbar setup script for real, since the already-
pinned "Boring.lnk" on this machine pointed at the old devs/taskbar
paths - verified its target now correctly points at
cases/devs/taskbar/run_clipsave.bat, and that the moved .bat actually
runs clipsave correctly against the real clipboard.

Added .gitattributes (`* text=auto eol=lf`) to fix the recurring
CRLF/LF bug for good, instead of catching it by hand on every edit.
Verified it works: an edit that produced CRLF locally got a "CRLF
will be replaced by LF" warning and staged correctly as LF with no
manual intervention. Ran `git add --renormalize .` to bring the whole
repo into line - caught two pre-existing CRLF files from before this
rule existed (core/configuration/user_conf.py, cases/webs/youtube.py)
and confirmed both changes are byte-for-byte identical content once
line endings are stripped, i.e. no actual code changed.

All 40 tests pass; every console-script entry point verified present
after `uv sync`.
