+++ cases/devs/taskbar/run_b64d.bat
@@ -1,6 +1,6 @@
 @echo off
-title Boring - clipsave
-"%~dp0..\..\.venv\Scripts\clipsave.exe"
+title Boring - b64d
+"%~dp0..\..\..\.venv\Scripts\b64d.exe"
 echo.
 echo Closing in 60 seconds - press any key to close now.
 timeout /t 60
