+++ cases/devs/taskbar/run_b64e.bat
@@ -1,6 +1,6 @@
 @echo off
 title Boring - b64e
-"%~dp0..\..\.venv\Scripts\b64e.exe"
+"%~dp0..\..\..\.venv\Scripts\b64e.exe"
 echo.
 echo Closing in 60 seconds - press any key to close now.
 timeout /t 60
