+++ cases/devs/taskbar/run_clipsave.bat
@@ -1,6 +1,6 @@
 @echo off
-title Boring - b64d
-"%~dp0..\..\.venv\Scripts\b64d.exe"
+title Boring - clipsave
+"%~dp0..\..\..\.venv\Scripts\clipsave.exe"
 echo.
 echo Closing in 60 seconds - press any key to close now.
 timeout /t 60
