+++ cases/wins/setBackgroundPicture.py
