+++ pyproject.toml
@@ -16,12 +16,12 @@ dependencies = [
 [project.scripts]
 hello = "scripts.hello:main"
 clean-logs = "scripts.cleanup:run"
-youtube = "scripts.webs.youtube:main"
-yt = "scripts.webs.youtube:main"
-clipsave = "wins.clipsave:main"
-b64d = "devs.base64_clip:main_decode"
-b64e = "devs.base64_clip:main_encode"
-prune-branches = "devs.prune_branches:main"
+youtube = "cases.webs.youtube:main"
+yt = "cases.webs.youtube:main"
+clipsave = "cases.wins.clipsave:main"
+b64d = "cases.devs.base64_clip:main_decode"
+b64e = "cases.devs.base64_clip:main_encode"
+prune-branches = "cases.devs.prune_branches:main"
 lucky = "cases.webs.lucky:main"
 mapit = "cases.maps.mapit:main"
 negative = "cases.pictures.negative:main"
